Update netomox (pseudo-dsl)
