+++ exe/convert_namespace/converter.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
+require 'netomox'
 require_relative 'convert_table'
 require_relative 'util'
-require_relative '../../model_defs/topology_builder/pseudo_dsl/pseudo_model'
 
 module TopologyOperator
   # namespace converter
@@ -18,7 +18,7 @@ module TopologyOperator
     # Rewrite all networks using convert table
     # @return [Hash] Converted topology data
     def convert
-      @dst_nws = TopologyBuilder::PseudoDSL::PNetworks.new
+      @dst_nws = Netomox::PseudoDSL::PNetworks.new
       @dst_nws.networks = @src_nws.networks
                                   .filter { |nw| target_network?(nw.name) }
                                   .map { |src_nw| rewrite_network(src_nw) }
@@ -61,9 +61,9 @@ module TopologyOperator
 
     # @param [Netomox::Topology::Node] src_node Source node (L3+)
     # @param [Netomox::Topology::TermPoint] src_tp Source term-point (L3+)
-    # @return [TopologyBuilder::PseudoDSL::PTermPoint]
+    # @return [Netomox::PseudoDSL::PTermPoint]
     def rewrite_term_point(src_node, src_tp)
-      dst_tp = TopologyBuilder::PseudoDSL::PTermPoint.new(convert_tp_name(src_node.name, src_tp.name))
+      dst_tp = Netomox::PseudoDSL::PTermPoint.new(convert_tp_name(src_node.name, src_tp.name))
       dst_tp.attribute = convert_all_hash_keys(src_tp.attribute.to_data)
       dst_tp.supports = rewrite_tp_supports(src_tp)
       dst_tp
@@ -84,9 +84,9 @@ module TopologyOperator
     end
 
     # @param [Netomox::Topology::Node] src_node Source node (L3+)
-    # @return [TopologyBuilder::PseudoDSL::PNode]
+    # @return [Netomox::PseudoDSL::PNode]
     def rewrite_node(src_node)
-      dst_node = TopologyBuilder::PseudoDSL::PNode.new(convert_node_name(src_node.name))
+      dst_node = Netomox::PseudoDSL::PNode.new(convert_node_name(src_node.name))
       dst_node.tps = src_node.termination_points.map { |src_tp| rewrite_term_point(src_node, src_tp) }
       dst_node.attribute = convert_all_hash_keys(src_node.attribute.to_data)
       dst_node.supports = rewrite_node_support(src_node)
@@ -94,27 +94,27 @@ module TopologyOperator
     end
 
     # @param [Netomox::Topology::TpRef] orig_edge Original link edge
-    # @return [TopologyBuilder::PseudoDSL::PLinkEdge]
+    # @return [Netomox::PseudoDSL::PLinkEdge]
     def rewrite_link_edge(orig_edge)
       node = convert_node_name(orig_edge.node_ref)
       tp = convert_tp_name(orig_edge.node_ref, orig_edge.tp_ref)
-      TopologyBuilder::PseudoDSL::PLinkEdge.new(node, tp)
+      Netomox::PseudoDSL::PLinkEdge.new(node, tp)
     end
 
     # @param [Netomox::Topology::Link] src_link Source link (L3+)
-    # @return [TopologyBuilder::PseudoDSL::PLink]
+    # @return [Netomox::PseudoDSL::PLink]
     def rewrite_link(src_link)
       dst_link_dst = rewrite_link_edge(src_link.source)
       dst_link_src = rewrite_link_edge(src_link.destination)
-      TopologyBuilder::PseudoDSL::PLink.new(dst_link_dst, dst_link_src)
+      Netomox::PseudoDSL::PLink.new(dst_link_dst, dst_link_src)
     end
 
     # rubocop:disable Metrics/AbcSize
 
     # @param [Netomox::Topology::Network] src_nw Source network (L3+)
-    # @return [TopologyBuilder::PseudoDSL::PNetwork]
+    # @return [Netomox::PseudoDSL::PNetwork]
     def rewrite_network(src_nw)
-      dst_nw = TopologyBuilder::PseudoDSL::PNetwork.new(src_nw.name)
+      dst_nw = Netomox::PseudoDSL::PNetwork.new(src_nw.name)
       # NOTE: network type is iterable hash
       dst_nw.type = src_nw.network_types.keys[0]
       dst_nw.attribute = convert_all_hash_keys(src_nw.attribute.to_data) if src_nw.attribute
