--- model_defs/topology_builder/pseudo_dsl/p_term_point.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'p_object_base'
-
-module TopologyBuilder
-  module PseudoDSL
-    # pseudo termination point
-    class PTermPoint < PObjectBase
-      # @return [String] String
-      def to_s
-        "[#{name}]"
-      end
-    end
-  end
-end
